+++ src/transports/Web/Web.ClientSupport.js
@@ -3,9 +3,9 @@
 
 const LIB_FS = require( 'fs' );
 const LIB_PATH = require( 'path' );
-const SRC_CLIENT_SUPPORT_CLIENT_API = require( './ClientSupport/Web.ClientSupport.ClientApiFile.js' );
-const SRC_CLIENT_SUPPORT_OPEN_API = require( './ClientSupport/Web.ClientSupport.OpenApiFile.js' );
-const SRC_CLIENT_SUPPORT_VIEW_CORE = require( './ClientSupport/Web.ClientSupport.ViewCore.js' );
+const SRC_CLIENT_SUPPORT_CLIENT_API = require( './Web.ClientSupport.ClientApiFile.js' );
+const SRC_CLIENT_SUPPORT_OPEN_API = require( './Web.ClientSupport.OpenApiFile.js' );
+const SRC_CLIENT_SUPPORT_VIEW_CORE = require( './Web.ClientSupport.ViewCore.js' );
 
 
 //---------------------------------------------------------------------
@@ -21,7 +21,7 @@ exports.ClientSupport_ViewEngine =
 			LIB_FS.mkdirSync( path, { recursive: true } );
 			CTX.Transport.ExpressApp.set( 'view engine', engine );
 			CTX.Transport.ExpressApp.set( 'views', path );
-			CTX.Server.Log.trace( `Web.ClientSupport is using '${engine}' views from folder [${CTX.Transport.Settings.ClientSupport.view_folder}].` );
+			CTX.Server.Log.trace( `Web.ClientSupport is using '${engine}' views from folder at [${CTX.Transport.Settings.ClientSupport.view_folder}].` );
 		}
 	};
 
@@ -40,7 +40,7 @@ exports.ClientSupport_GenerateClientApiFile =
 			let filename = CTX.Server.ResolveApplicationPath( CTX.Transport.Settings.ClientSupport.client_api_file );
 			LIB_FS.mkdirSync( LIB_PATH.dirname( filename ), { recursive: true } );
 			LIB_FS.writeFileSync( filename, code );
-			CTX.Server.Log.trace( `Web.ClientSupport generated client file [${CTX.Transport.Settings.ClientSupport.client_api_file}].` );
+			CTX.Server.Log.trace( `Web.ClientSupport generated client file at [${CTX.Transport.Settings.ClientSupport.client_api_file}].` );
 		}
 	};
 
@@ -59,7 +59,7 @@ exports.ClientSupport_GenerateOpenApiFile =
 			let filename = CTX.Server.ResolveApplicationPath( CTX.Transport.Settings.ClientSupport.open_api_file );
 			LIB_FS.mkdirSync( LIB_PATH.dirname( filename ), { recursive: true } );
 			LIB_FS.writeFileSync( filename, JSON.stringify( code ) );
-			CTX.Server.Log.trace( `Web.ClientSupport generated open api file [${CTX.Transport.Settings.ClientSupport.open_api_file}].` );
+			CTX.Server.Log.trace( `Web.ClientSupport generated open api file at [${CTX.Transport.Settings.ClientSupport.open_api_file}].` );
 		}
 	};
 
@@ -75,7 +75,7 @@ exports.ClientSupport_MountPublicFolder =
 			let public_folder = CTX.Server.ResolveApplicationPath( CTX.Transport.Settings.ClientSupport.public_folder );
 			LIB_FS.mkdirSync( public_folder, { recursive: true } );
 			CTX.Transport.ExpressApp.use( public_url, CTX.LIB_EXPRESS.static( public_folder ) );
-			CTX.Server.Log.trace( `Web.ClientSupport mounted route [${public_url}] for public folder [${CTX.Transport.Settings.ClientSupport.public_folder}].` );
+			CTX.Server.Log.trace( `Web.ClientSupport mounted route [${public_url}] for public folder at [${CTX.Transport.Settings.ClientSupport.public_folder}].` );
 		}
 	};
 
@@ -91,10 +91,12 @@ exports.ClientSupport_MountRootRoute =
 	{
 		if ( CTX.Transport.Settings.ClientSupport
 			&& CTX.Transport.Settings.ClientSupport.enabled
-			&& CTX.Transport.Settings.ClientSupport.Views.root_view )
+			// && CTX.Transport.Settings.ClientSupport.Views.root_view )
+			&& CTX.Transport.Settings.ClientSupport.home_view )
 		{
 			let root_url = CTX.Transport.ServerPath();
-			let root_view = CTX.Transport.Settings.ClientSupport.Views.root_view;
+			// let root_view = CTX.Transport.Settings.ClientSupport.Views.root_view;
+			let root_view = CTX.Transport.Settings.ClientSupport.home_view;
 			CTX.Transport.ExpressApp.get( root_url,
 				// CTX.Transport.AuthenticationGate( false ),
 				// CTX.Transport.AuthorizationGate( [ '*' ] ),